Should be in a header file
