+++ WangscapeTest/TestRequiringOptions.cpp
@@ -1,19 +1,8 @@
-#include <gtest/gtest.h>
+#include "TestRequiringOptions.h"
 
-#include <Options.h>
-#include <OptionsManager.h>
-
-class TestRequiringOptions : public ::testing::Test {
-protected:
-    std::string filename;
-    const Options& options;
-    const OptionsManager optionsManager;
-
-    TestRequiringOptions() :
-        filename("../../Wangscape/example/example_options.json"),
-        optionsManager(filename),
-        options(optionsManager.getOptions())
-    {
-    };
-   virtual ~TestRequiringOptions() = default;
-};
+TestRequiringOptions::TestRequiringOptions() :
+    filename("../../Wangscape/example/example_options.json"),
+    optionsManager(filename),
+    options(optionsManager.getOptions())
+{
+}
