+++ WangscapeTest/TestRequiringOptions.h
@@ -1,0 +1,14 @@
+#include <gtest/gtest.h>
+
+#include <Options.h>
+#include <OptionsManager.h>
+
+class TestRequiringOptions : public ::testing::Test {
+protected:
+    std::string filename;
+    const Options& options;
+    const OptionsManager optionsManager;
+
+    TestRequiringOptions();
+    virtual ~TestRequiringOptions() = default;
+};
